search뷰컨 검색기록 저장하는 함수 리팩토링
