+++ INUBus/Cell/SearchTableViewCell.swift
@@ -28,6 +28,7 @@ class SearchTableViewCell: UITableViewCell {
     // Configure the view for the selected state
   }
   
+  
   //deletebutton을 누르면 검색기록의 해당 row의 값을 지워줌
   @IBAction func deleteButtonDidTap() {
